requirements.txt
tensorflow>=2.12.0
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
jupyter>=1.0.0
scipy>=1.11.0
